Перенесены вью и сериализаторы в api/v1, добавлены урлы на комменты и отзывы
