--- api_yamdb/api/serializers.py
@@ -1,32 +1,0 @@
-from rest_framework import serializers
-
-from rest_framework.validators import UniqueTogetherValidator
-
-from reviews.models import Comment, Review
-
-class ReviewSerializer(serializers.ModelSerializer):
-    author = serializers.SlugRelatedField(
-        read_only=True,
-        slug_field='username'
-    )
-
-    class Meta:
-        model = Review
-        fields = ['id', 'text', 'author', 'score', 'pub_date']
-        validators = [
-            UniqueTogetherValidator(
-                queryset=Review.objects.all(),
-                fields=('author', 'title',)
-            )
-        ]
-
-
-class CommentSerializer(serializers.ModelSerializer):
-    author = serializers.SlugRelatedField(
-        read_only=True,
-        slug_field='username'
-    )
-
-    class Meta:
-        model = Comment
-        fields = ['id', 'text', 'author', 'pub_date']
