+++ api_yamdb/api/v1/urls.py
@@ -2,10 +2,22 @@ from django.urls import include, path
 
 from rest_framework import routers
 
-from .views import GenreViewSet, CategoryViewSet, TitleViewSet
+from .views import (CategoryViewSet, CommentViewSet,
+                    GenreViewSet, ReviewViewSet,
+                    TitleViewSet)
 
 router = routers.DefaultRouter()
 router.register(r'titles', TitleViewSet)
+router.register(
+    r'titles/(?P<title_id>\d+)/reviews',
+    ReviewViewSet,
+    basename='reviews'
+)
+router.register(
+    r'titles/(?P<title_id>\d+)/reviews/(?P<review_id>\d+)/comments',
+    CommentViewSet,
+    basename='comments'
+)
 
 urlpatterns = [
     path('', include(router.urls)),
