+++ api_yamdb/api/v1/views.py
@@ -2,12 +2,12 @@ from django_filters.rest_framework import DjangoFilterBackend
 
 from django.shortcuts import get_object_or_404
 
-from rest_framework import viewsets, mixins
+from rest_framework import mixins, viewsets
 
-from reviews.models import Genre, Category, Title
-from .serializers import (
-    GenreSerializer, CategorySerializer, TitleSerializer, TitlePostSerializer
-)
+from reviews.models import Category, Genre, Title, Review
+from .serializers import (CategorySerializer,CommentSerializer,
+                          GenreSerializer, TitleSerializer,
+                          TitlePostSerializer, ReviewSerializer)
 
 
 class GenreViewSet(mixins.CreateModelMixin,
@@ -48,3 +48,31 @@ class TitleViewSet(viewsets.ModelViewSet):
         if self.action in list:
             return TitlePostSerializer
         return TitleSerializer
+
+
+class ReviewViewSet(viewsets.ModelViewSet):
+    """Вьюсет для модели Review."""
+    serializer_class = ReviewSerializer
+    #permission_classes = []
+
+    def perform_create(self, serializer):
+        title = get_object_or_404(Title, id=self.kwargs.get('title_id'))
+        serializer.save(author=self.request.user, title=title)
+
+    def get_queryset(self):
+        title = get_object_or_404(Title, id=self.kwargs.get('title_id'))
+        return title.reviews.all()
+
+
+class CommentViewSet(viewsets.ModelViewSet):
+    """Вьюсет для модели Comment."""
+    serializer_class = CommentSerializer
+    #permission_classes = []
+
+    def perform_create(self, serializer):
+        review = get_object_or_404(Review, id=self.kwargs.get('review_id'))
+        serializer.save(author=self.request.user, review=review)
+
+    def get_queryset(self):
+        review = get_object_or_404(Review, id=self.kwargs.get('review_id'))
+        return review.comments.all()
